requirements.txt
pymongo==4.16.0
requests==2.32.3
pdfplumber==0.11.9
python-dotenv==1.2.2
